+++ filen-cli/CHANGELOG.md
@@ -12,6 +12,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - rclone commands (`mount`, `serve`) are more customizable via `--cache-size` and `--transfers` flags
   as well as the option to take any number of custom Rclone options
 - convenient install script on Linux and macOS
+- convenient script to temporarily install the CLI and execute `export-api-key` for use with Rclone
+
+### Changed
+
+- unified `webdav`, `ftp`, `sftp`, `http-server`, `s3` commands as `serve` command
+
+### Fixed
+
+- `cd` command: check if directory is valid #7
+- `serve` command: don't automatically add IP to `--addr` option
 
 ## 0.2.4 - 2026-01-20