+++ filen-cli/CHANGELOG.md
@@ -5,9 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased 
+## Unreleased
 
 - /
+
+## 0.2.6 - 2026-06-04
+
+### Changed
+
+- managed Rclone now uses upstream Rclone (instead of filen-rclone fork), which is more maintained
 
 ## 0.2.5 - 2026-01-29